pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "gtav-slms"
version = "0.1.0"
description = "Small Language Models for GTA V gameplay automation"
readme = "README.md"
license = {text = "Apache-2.0"}
requires-python = ">=3.10"
authors = [
    {name = "Shrijeeth"},
]
keywords = ["gta5", "game-ai", "imitation-learning", "slm", "vision-model"]
classifiers = [
    "Development Status :: 2 - Pre-Alpha",
    "Programming Language :: Python :: 3",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "torch>=2.0",
    "torchvision>=0.15",
    "timm>=0.9",
    "numpy>=1.24",
    "opencv-python>=4.8",
    "Pillow>=10.0",
    "pyyaml>=6.0",
    "pynput>=1.7",
    "mss>=9.0",
    "h5py>=3.9",
    "zarr>=2.16",
    "tqdm>=4.65",
    "wandb>=0.15",
    "click>=8.1",
    "onnxruntime>=1.16",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "black>=23.0",
    "ruff>=0.1",
]

[project.scripts]
gtav-slm = "main:cli"

[tool.setuptools.packages.find]
where = ["."]
include = ["models*", "envs*", "data*", "training*", "inference*", "utils*", "configs*"]
